+++ src/main/java/com/example/demo/domain/exceptions/RealityNotFoundException.java
@@ -1,4 +1,4 @@
-package com.example.demo.reality;
+package com.example.demo.domain.exceptions;
 
 public class RealityNotFoundException extends Exception {
     public RealityNotFoundException(String message) {
